fog/requirements.txt
numpy>=1.26.0,<2.0.0
opencv-python>=4.10.0
boto3>=1.34.0
botocore>=1.34.0
awsiotsdk>=1.12.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
pyyaml==6.0.1

# ML deps (CPU-only torch)
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple
torch==2.5.0+cpu
torchvision==0.20.0+cpu
ultralytics>=8.3.0
